Change the condition variable for automatic update
Use a dictionary (g_parameters) instead of a boolean (UPDATE_AT_START_P)
to determine whether automatic update is enabled or not.  This behavior
should be specified via request parameters, and more behavior like this
may be added for the future.  So the condition variable should be
a dictionary for further extensibility.