Use @class instead of @id to record tweet id
There is a planned feature such as multiple tabpages, so that a tweet
can be appeared in more than one place.  Currently, ID of a tweet is
recorded in @id, but it's not easy to update the appearances of all
"views" of a tweet, so that use @class instead of @id to record tweet
id.