Fix favorite_symbol() to return "raw" character
Old favorite_symbol() returns a symbol written in character entity
reference such as '&#x2605;', but it doesn't work for toggle_favorite()
because the interpretation of character entity reference is not done for
all cases.

Fix this bug by returning a raw character such as '\u2605' instead of
character entity reference.