Refactor - Use .data() instead of title attribute
Attribute "title" is attached to columns and column selectors to
identity each column.  But many browsers show a tooltip to indicate the
value of title attribute when the cursor hovers the corresponding
element.  This behavior is annoying, because this title attribute is for
internal use, not for users.

Fix this problem by using jQuery.data() instead of title attribute.