Simplify requesting Twitter API with jQuery.ajax()
There are two problems in process_queued_api_request_with_oauth():

(a) It uses $('#request_iframe').load() to call callback function for
success and this method works fine on Safari 5.0 (5533.16), but it does
not work for Opera 10.  And onload attribute for <iframe> is not defined
in both HTML 4.01 Frameset and XHTML 1.0 Frameset.

(b) The way to detect errors from a request is not reliable.

After several tests, I confirmed that it's possible to use jQuery.ajax()
to send a request.  Though there is a problem that jQuery.ajax() may
alter a given "url" parameter, there is also workaround for the problem.