merge_tweets_n2o: Assure to drop existing tweets
It seems not to be necessary to drop existing tweets in merge_tweets_n2o
and/or queue_tweets_n2o, so that such filtering was removed before.  But
sometimes Twitter returns duplicate tweets even if since_id is
specified.

Change merge_tweets_n2o to assure to drop existing tweets to avoid
unexpected result.