% Test; Use jquery.hotkeys.js
jquery.hotkeys.js 0.7.9 doesn't work properly with jQuery 1.4.2.  Event
handler for a hotkey will be called two or more times.  The problem is
already known and it's currently discussed but it's not fixed yet.  See:
http://code.google.com/p/js-hotkeys/issues/detail?id=78

jquery.hotkeys.js 0.7.9 doesn't work properly with jQuery 1.4.4.  Event
handler for a hotkey will never be called.  The problem is also known
and it's not fixed yet.  See:
http://code.google.com/p/js-hotkeys/issues/detail?id=91

Therefore, hotkey feature is pending until the above problems are fixed.