Preference.apply: Do not save value if possible
Preference.save (actually jQuery.toJSON) takes a long time if
a preference value is large enough.  And Preference.apply always gets
new value from its form and saves the value.  But most of preferences
are not changed whenever user clicks the "Apply" button.

Change Preference.apply not to save value if the value is not modified.